# Small additional speedup for deflate_quick